Django_Assignment/requirements.txt
Django==2.1.11
django-crispy-forms==1.7.2
Pillow==6.2.0
pytz==2018.4
djangorestframework==3.11.0
